+++ index.mjs
@@ -1,0 +1,4 @@
+import hotShots from './index.js';
+
+export const StatsD = hotShots.StatsD;
+export default hotShots;